// src/app/models/employee.model.ts
export interface Employee {
    id: string;
    accountId: number;
    accountStatus: string;
    businessAccountId: string;
    firstname: string;
    lastname: string;
    fullname: string;
    role: string;
    type: string;
    address: string;
    city: string;
    state: string;
    country: string;
    zipcode: string;
    email: string;
    phoneNumber: string;
    userId: string;
    createdAt: string;
    updatedAt: string;
}